add stderr redirect to stdout on adduser
When SecurityOnion is installed, but not yet configured, on a base Ubuntu 12.04.5 server install, it creates the watchdog cron.d entry which is called every five minutes.  However, the sguil user has not yet been created, so nsm_sensor_ps-restart calls adduser with an invalid username, and the error is sent to stderr, which generates an email to the default user every five minutes.
By redirecting stderr to stdout (before redirecting stdout to /dev/null), when adduser is called before sosetup is run (sguil user does not yet exist), the adduser error shows up in /var/log/nsm/watchdog.log rather than being mailed to root.